FIxed middleware bundle size
