+++ src/auth.ts
@@ -1,59 +1,5 @@
-import NextAuth, { DefaultSession } from "next-auth";
-import { Provider } from "next-auth/providers";
-import GitHub from "next-auth/providers/github";
-import { PrismaAdapter } from "@auth/prisma-adapter";
-import { PrismaClient, Role } from "../generated/prisma";
-import { PoolConfig } from "@neondatabase/serverless";
-import { PrismaNeon } from "@prisma/adapter-neon";
-
-const neon: PoolConfig = {
-  connectionString: process.env.AUTH_POSTGRES_PRISMA_URL,
-};
-const adapter = new PrismaNeon(neon);
-const prisma = new PrismaClient({ adapter });
-
-declare module "next-auth" {
-  /**
-   * The shape of the user object returned in the OAuth providers' `profile` callback,
-   * or the second parameter of the `session` callback, when using a database.
-   */
-  interface User {
-    role: Role;
-  }
-  /**
-   * The shape of the account object returned in the OAuth providers' `account` callback,
-   * Usually contains information about the provider being used, like OAuth tokens (`access_token`, etc).
-   */
-  // interface Account {}
-
-  /**
-   * Returned by `useSession`, `auth`, contains information about the active session.
-   */
-  interface Session {
-    user: {
-      role: string;
-    } & DefaultSession["user"];
-  }
-}
-
-const providers: Provider[] = [
-  GitHub({
-    clientId: process.env.AUTH_GITHUB_ID,
-    clientSecret: process.env.AUTH_GITHUB_SECRET,
-    allowDangerousEmailAccountLinking: true,
-  }),
-];
-
-export const providerMap = providers
-  .map((provider) => {
-    if (typeof provider === "function") {
-      const providerData = provider();
-      return { id: providerData.id, name: providerData.name };
-    } else {
-      return { id: provider.id, name: provider.name };
-    }
-  })
-  .filter((provider) => provider.id !== "credentials");
+import NextAuth from "next-auth";
+import authConfig from "./auth.config";
 
 export const {
   handlers,
@@ -62,13 +8,10 @@ export const {
   signOut,
   unstable_update: update,
 } = NextAuth({
-  adapter: PrismaAdapter(prisma),
-
+  ...authConfig,
   // pages: { signIn: "/signin" },
-  providers: providers,
   session: { strategy: "database" },
   callbacks: {
-
     session({ session, user }) {
       session.user.role = user.role;
       return session;
